services/azure_functions/requirements.txt
azure-functions
azure-iot-hub